LayoutTests/plugins/plugin-initiate-popup-window.html

<html>
<head>
<script>
if (window.layoutTestController) {
  layoutTestController.dumpAsText();
  layoutTestController.setCanOpenWindows();
  layoutTestController.setPopupBlockingEnabled(true);
  layoutTestController.setCloseRemainingWindowsWhenComplete(true);
  layoutTestController.waitUntilDone();
}

function test() {
    if (window.layoutTestController) {
        // Send a mouse-click event to set the input focus to the test plug-in.
        eventSender.mouseMoveTo(0, 0);
        eventSender.mouseMoveTo(20, 20);
        eventSender.mouseDown();
        eventSender.mouseUp();
    }
}

function popup_by_mousedown() {
    // We rely on sequential ordering of POST processing.
    var form = document.getElementById('form');
    form.submit();
    setTimeout(send_key_to_plugin, 1);
}

function send_key_to_plugin() {
  // Send keyboard events to the plug-in.
  if (window.layoutTestController) {
      eventSender.mouseMoveTo(0, 0);
      eventSender.mouseMoveTo(60, 60);
      eventSender.mouseDown();
      eventSender.mouseUp();
      eventSender.keyDown('p');
  }
}

function popup_by_keydown() {
    // We rely on sequential ordering of POST processing.
    var form = document.getElementById('form2');
    form.submit();
    setTimeout(close_test, 1);
}
function close_test() {
    var form = document.getElementById('form3');
    form.submit();
}
</script>
</head>
<body onload="test()">
<embed name="plg" type="application/x-webkit-test-netscape" width=100 height=40 evaluatescript="mouse::popup_by_mousedown()"></embed><br>
<embed type="application/x-webkit-test-netscape" width=100 height=40 evaluatescript="key::popup_by_keydown()"></embed><br>
Specify a script and a mouse/keyboard event to the plugin. The specified script will be evaluated in the browser when the specified event is received by the plugin. The test is for bug https://bugs.webkit.org/show_bug.cgi?id=41292.
<form id="form" action="data:text/html,<script>alert('MOUSEDOWN PASSED')</script>" target="target1">
<input type="submit" value="Submit to new window"/>
</form>
<form id="form2" action="data:text/html,<script>alert('KEYDOWN PASSED')</script>" target="target2">
<input type="submit" value="Submit to new window"/>
</form>
<form id="form3" action="data:text/html,<b>hello!</b><script>window.layoutTestController && layoutTestController.notifyDone()</script>" target="panel">
<input type="submit" value="Submit local page Javascript"/>
</form>
<iframe id="target1"></iframe><iframe id="target2"></iframe><iframe id="panel"></iframe>
</body>
</html>
