LayoutTests/plugins/plugin-initiate-popup-window-expected.txt

CONSOLE MESSAGE: line 0: PLUGIN: getFocusEvent
CONSOLE MESSAGE: line 0: PLUGIN: getFocusEvent
CONSOLE MESSAGE: line 0: PLUGIN: getFocusEvent
CONSOLE MESSAGE: line 0: PLUGIN: mouseDown at (12, 12)
CONSOLE MESSAGE: line 0: PLUGIN: mouseUp at (12, 12)
CONSOLE MESSAGE: line 0: PLUGIN: loseFocusEvent
CONSOLE MESSAGE: line 0: PLUGIN: getFocusEvent
CONSOLE MESSAGE: line 0: PLUGIN: mouseDown at (52, 12)
CONSOLE MESSAGE: line 0: PLUGIN: mouseUp at (52, 12)
CONSOLE MESSAGE: line 0: PLUGIN: keyDown 'p'
CONSOLE MESSAGE: line 0: PLUGIN: keyUp 'p'
CONSOLE MESSAGE: line 0: PLUGIN: loseFocusEvent
ALERT: MOUSEDOWN PASSED
ALERT: KEYDOWN PASSED


Specify a script and a mouse/keyboard event to the plugin. The specified script will be evaluated in the browser when the specified event is received by the plugin. The test is for bug https://bugs.webkit.org/show_bug.cgi?id=41292.
